add signupcomponent
