+++ Frontend/src/app/app.component.html
@@ -1,2 +1,3 @@
 <app-header></app-header>
+
 <router-outlet></router-outlet>
