+++ Frontend/src/app/app.config.ts
@@ -5,12 +5,18 @@ import { routes } from './app.routes';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 import { authInterceptor } from './services/auth.interceptor';
+import { provideAnimations } from '@angular/platform-browser/animations';//update
+import { importProvidersFrom } from '@angular/core';//update
+import { ReactiveFormsModule } from '@angular/forms';//update
+import { CommonModule } from '@angular/common';//update
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes, withComponentInputBinding(), withViewTransitions()),
     provideHttpClient(withInterceptors([authInterceptor])),
     provideAnimationsAsync(),
+    provideAnimations(),//update
+    importProvidersFrom(ReactiveFormsModule,CommonModule)//update
     
   ]
 };
