--- Frontend/src/app/components/login/login.component.html
@@ -1,55 +1,0 @@
-<div class="flex justify-center items-center min-h-screen bg-gray-100">
-          <div class="bg-white p-8 rounded-lg shadow-md max-w-sm w-full">
-            
-            
-            <h2 class="text-2xl font-bold text-center text-gray-800 mb-4">Login</h2>
-        
-           
-            <form [formGroup]="loginForm" (ngSubmit)="login()" class="flex flex-col gap-4">
-              
-              
-              <label class="block">
-                <span class="text-gray-700">Email</span>
-                <input type="email" formControlName="email" placeholder="Enter your email" 
-                  class="mt-1 p-2 w-full border rounded-lg focus:ring-2 focus:ring-blue-300" required />
-                
-                <div *ngIf="loginForm.get('email')?.invalid && loginForm.get('email')?.touched" class="text-red-500 text-sm mt-1">
-                  <small *ngIf="loginForm.get('email')?.errors?.['required']">Email is required.</small>
-                  <small *ngIf="loginForm.get('email')?.errors?.['email']">Invalid email format.</small>
-                </div>
-              </label>
-        
-              
-              <label class="block">
-                <span class="text-gray-700">Password</span>
-                <input type="password" formControlName="password" placeholder="Enter your password" 
-                  class="mt-1 p-2 w-full border rounded-lg focus:ring-2 focus:ring-blue-300" required />
-                
-                <div *ngIf="loginForm.get('password')?.invalid && loginForm.get('password')?.touched" class="text-red-500 text-sm mt-1">
-                  <small *ngIf="loginForm.get('password')?.errors?.['required']">Password is required.</small>
-                  <small *ngIf="loginForm.get('password')?.errors?.['minlength']">Password must be at least 6 characters.</small>
-                  <small *ngIf="loginForm.get('password')?.errors?.['pattern']">Password must include uppercase, lowercase, and a number.</small>
-                </div>
-              </label>
-        
-              
-              <p *ngIf="errorMessage" class="text-red-500 text-sm text-center">{{ errorMessage }}</p>
-        
-              
-          <button type="submit" [disabled]="loginForm.invalid || isLoading"
-              class="w-full bg-blue-500 hover:bg-blue-600 text-white font-medium py-2 rounded-lg transition
-               disabled:bg-gray-400 disabled:cursor-not-allowed">
-        <span *ngIf="isLoading">Loading...</span>
-        <span *ngIf="!isLoading">Login</span>
-      </button>
-        
-              
-              <p class="text-sm text-gray-600 text-center mt-2">
-                Don't have an account? 
-                <a routerLink="/register" class="text-blue-500 hover:underline">Sign up</a>
-              </p>
-        
-            </form>
-          </div>
-        </div>
-
