--- Frontend/src/app/components/register/register.component.ts
@@ -1,83 +1,0 @@
-import { Component } from '@angular/core';
-import { Router } from '@angular/router';
-import { FormBuilder, FormGroup, Validators, AbstractControl, ValidationErrors, ReactiveFormsModule } from '@angular/forms';
-import { AuthService } from '../../services/auth.service';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'app-register',
-  standalone: true,
-  imports: [CommonModule, ReactiveFormsModule],
-  templateUrl: './register.component.html',
-  styleUrls: ['./register.component.css'],
-})
-export class RegisterComponent {
-  registerForm: FormGroup;
-  errorMessage: string = '';
-  isLoading: boolean = false;
-  isPasswordVisible: boolean = false;
-  isConfirmPasswordVisible: boolean = false;
-  formSubmitted: boolean = false;
-  
-
-  constructor(
-    private fb: FormBuilder,
-    private authService: AuthService,
-    private router: Router
-  ) {
-    this.registerForm = this.fb.group(
-      {
-        name: ['', [Validators.required, Validators.pattern(/^[A-Za-z\s]+$/)]],
-        email: ['', [Validators.required, Validators.email]],
-        password: ['', [
-          Validators.required, 
-          Validators.minLength(8), 
-          Validators.pattern(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).{8,}$/)]],
-        confirmPassword: ['', Validators.required],
-      },
-      { validator: this.passwordMatchValidator }
-    );
-  }
-
-  register() {
-    this.formSubmitted = true;
-
-    if (this.registerForm.invalid){
-      
-      return;
-    }
-    this.errorMessage = '';
-    this.isLoading = true;
-
-
-    const { name, email, password } = this.registerForm.value;
-
-    this.authService.register(name, email, password, 'Customer').subscribe({
-      next: () => {
-        this.isLoading = false;
-        this.router.navigate(['/login']);
-      },
-      error: (error) => {
-        this.isLoading = false;
-        this.errorMessage = error?.error?.message || 'Registration failed!';
-      },
-    });
-  }
-
- 
-  passwordMatchValidator(control: AbstractControl): ValidationErrors | null {
-    const password = control.get('password')?.value;
-    const confirmPassword = control.get('confirmPassword')?.value;
-    return password === confirmPassword ? null : { passwordNotMatch: true };
-  }
-
-  togglePasswordVisibility() {
-    this.isPasswordVisible = !this.isPasswordVisible;
-  }
-  
-  toggleConfirmPasswordVisibility() {
-    this.isConfirmPasswordVisible = !this.isConfirmPasswordVisible;
-  }
-
-  
-}
