+++ Frontend/src/styles.css
@@ -1,2 +1,5 @@
 /* You can add global styles to this file, and also import other style files */
-@import "tailwindcss";
+ @import "tailwindcss"; 
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
